CrimesAnalysis_GCP/FINAL-ARCHI1/source/requirements.txt
google-api-python-client==2.52.0
google-auth==2.6.0
google-cloud-storage==2.3.0
